Make it more extendable...
