+++ Source/PropertyTools.Wpf/DataGrid/Definitions/TemplateColumnDefinition.cs
@@ -7,6 +7,8 @@
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
+using System.Windows.Controls;
+
 namespace PropertyTools.Wpf
 {
     using System.Windows;
@@ -33,5 +35,47 @@ namespace PropertyTools.Wpf
         /// </summary>
         /// <value>The CellEditingTemplate.</value>
         public DataTemplate CellEditingTemplate { get; set; }
+
+        /// <summary>
+        /// The Default DataGridControlFactory uses this Method to Create the Control
+        /// </summary>
+        /// <param name="bindingPath"></param>
+        /// <returns></returns>
+        public override FrameworkElement CreateDisplayControl(string bindingPath)
+        {
+            var template = this.CellTemplate;
+            var element = template.LoadContent() as FrameworkElement;
+            var binding = this.CreateBinding(bindingPath);
+            var contentControl = new ContentControl()
+            {
+                HorizontalAlignment = HorizontalAlignment.Stretch,
+                VerticalAlignment = VerticalAlignment.Stretch,
+                Content = element,
+            };
+            element.SetBinding(FrameworkElement.DataContextProperty, binding);
+            return contentControl;
+        }
+
+        /// <summary>
+        /// The Default DataGridControlFactory uses this Method to Create the Edit Control
+        /// </summary>
+        /// <param name="bindingPath"></param>
+        /// <returns></returns>
+        public override FrameworkElement CreateEditControl(string bindingPath)
+        {
+            var template = this.CellEditingTemplate;
+            if (template == null)
+                return this.CreateDisplayControl(bindingPath);
+            var element = template.LoadContent() as FrameworkElement;
+            var binding = this.CreateBinding(bindingPath);
+            var contentControl = new ContentControl()
+            {
+                HorizontalAlignment = HorizontalAlignment.Stretch,
+                VerticalAlignment = VerticalAlignment.Stretch,
+                Content = element,
+            };
+            element.SetBinding(FrameworkElement.DataContextProperty, binding);
+            return contentControl;
+        }
     }
 }
